+++ src/collision/g3dlite/G3D/Vector4.inl
@@ -9,7 +9,7 @@
 
 //----------------------------------------------------------------------------
 
-inline unsigned int hashCode(const G3D::Vector4& v) {
+inline size_t hashCode(const G3D::Vector4& v) {
      return v.hashCode();
 }
