+++ src/collision/g3dlite/Vector3.cpp
@@ -63,10 +63,10 @@ Vector3::Axis Vector3::primaryAxis() const {
 }
 
 
-unsigned int Vector3::hashCode() const {
-    unsigned int xhash = (*(int*)(void*)(&x));
-    unsigned int yhash = (*(int*)(void*)(&y));
-    unsigned int zhash = (*(int*)(void*)(&z));
+size_t Vector3::hashCode() const {
+    size_t xhash = (*(size_t*)(void*)(&x));
+    size_t yhash = (*(size_t*)(void*)(&y));
+    size_t zhash = (*(size_t*)(void*)(&z));
 
     return xhash + (yhash * 37) + (zhash * 101);
 }
